modify error version 6
