+++ examples/experimental/control_net_canny.py
@@ -23,6 +23,11 @@ image = np.array(image)
 
 LOW_THRESHOLD = 100
 HIGH_THRESHOLD = 200
+PROMPT = "disco dancer with colorful lights, best quality, extremely detailed"
+NEGATIVE_PROMPT = (
+    "longbody, lowres, bad anatomy, bad hands, "
+    " missing fingers, extra digit, fewer digits, cropped, worst quality, low quality"
+)
 
 image = cv2.Canny(image, LOW_THRESHOLD, HIGH_THRESHOLD)
 image = image[:, :, None]
@@ -41,11 +46,6 @@ pipe.to("cuda")
 pipe.unet = oneflow_compile(pipe.unet)
 generator = torch.manual_seed(0)
 
-PROMPT = "disco dancer with colorful lights, best quality, extremely detailed"
-NEGATIVE_PROMPT = (
-    "longbody, lowres, bad anatomy, bad hands, "
-    " missing fingers, extra digit, fewer digits, cropped, worst quality, low quality"
-)
 
 out_images = pipe(
     prompt=PROMPT,
