+++ examples/experimental/text_to_image_sdxl_fp16.py
@@ -9,7 +9,6 @@ from diffusers import StableDiffusionXLPipeline
 import torch
 
 from onediff.infer_compiler import torchbackend
-from onediff.infer_compiler.utils.set_oneflow_environment import set_oneflow_environment
 
 parser = argparse.ArgumentParser()
 parser.add_argument(
@@ -45,7 +44,20 @@ pipe = StableDiffusionXLPipeline.from_pretrained(
 if args.compile:
     if args.graph:
         os.environ["with_graph"] = "1"
-        set_oneflow_environment()
+        os.environ["ONEFLOW_MLIR_CSE"] = "1"
+        os.environ["ONEFLOW_MLIR_ENABLE_INFERENCE_OPTIMIZATION"] = "1"
+        os.environ["ONEFLOW_MLIR_ENABLE_ROUND_TRIP"] = "1"
+        os.environ["ONEFLOW_MLIR_FUSE_FORWARD_OPS"] = "1"
+        os.environ["ONEFLOW_MLIR_FUSE_OPS_WITH_BACKWARD_IMPL"] = "1"
+        os.environ["ONEFLOW_MLIR_GROUP_MATMUL"] = "1"
+        os.environ["ONEFLOW_MLIR_PREFER_NHWC"] = "1"
+        os.environ["ONEFLOW_KERNEL_ENABLE_FUSED_CONV_BIAS"] = "1"
+        os.environ["ONEFLOW_KERNEL_ENABLE_FUSED_LINEAR"] = "1"
+        os.environ["ONEFLOW_KERNEL_CONV_CUTLASS_IMPL_ENABLE_TUNING_WARMUP"] = "1"
+        os.environ["ONEFLOW_KERNEL_CONV_ENABLE_CUTLASS_IMPL"] = "1"
+        os.environ["ONEFLOW_CONV_ALLOW_HALF_PRECISION_ACCUMULATION"] = "1"
+        os.environ["ONEFLOW_MATMUL_ALLOW_HALF_PRECISION_ACCUMULATION"] = "1"
+        os.environ["ONEFLOW_LINEAR_EMBEDDING_SKIP_INIT"] = "1"
     pipe.unet = torch.compile(
         pipe.unet, fullgraph=True, mode="reduce-overhead", backend=torchbackend
     )
