+++ src/infer_compiler_registry/register_diffusers_quant/quant_diffusion_pipeline.py
@@ -8,9 +8,33 @@ from diffusers_quant.utils import (
     replace_sub_module_with_quantizable_module,
 )
 from torch._dynamo import allow_in_graph as maybe_allow_in_graph
-from onediff.infer_compiler.utils.set_oneflow_environment import _use_graph
 
 
+def _use_graph():
+    os.environ["with_graph"] = "1"
+    os.environ["ONEFLOW_GRAPH_DELAY_VARIABLE_OP_EXECUTION"] = "1"
+    os.environ["ONEFLOW_MLIR_CSE"] = "1"
+    os.environ["ONEFLOW_MLIR_ENABLE_INFERENCE_OPTIMIZATION"] = "1"
+    os.environ["ONEFLOW_MLIR_ENABLE_ROUND_TRIP"] = "1"
+    os.environ["ONEFLOW_MLIR_FUSE_FORWARD_OPS"] = "1"
+    os.environ["ONEFLOW_MLIR_FUSE_OPS_WITH_BACKWARD_IMPL"] = "1"
+    os.environ["ONEFLOW_MLIR_GROUP_MATMUL"] = "1"
+    os.environ["ONEFLOW_MLIR_PREFER_NHWC"] = "1"
+    os.environ["ONEFLOW_KERNEL_ENABLE_FUSED_CONV_BIAS"] = "1"
+    os.environ["ONEFLOW_KERNEL_ENABLE_FUSED_LINEAR"] = "1"
+    os.environ["ONEFLOW_KERNEL_CONV_CUTLASS_IMPL_ENABLE_TUNING_WARMUP"] = "1"
+    os.environ["ONEFLOW_KERNEL_CONV_ENABLE_CUTLASS_IMPL"] = "1"
+    os.environ["ONEFLOW_KERNEL_GEMM_CUTLASS_IMPL_ENABLE_TUNING_WARMUP"] = "1"
+    os.environ["ONEFLOW_KERNEL_GEMM_ENABLE_CUTLASS_IMPL"] = "1"
+    os.environ["ONEFLOW_CONV_ALLOW_HALF_PRECISION_ACCUMULATION"] = "1"
+    os.environ["ONEFLOW_MATMUL_ALLOW_HALF_PRECISION_ACCUMULATION"] = "1"
+    os.environ["ONEFLOW_LINEAR_EMBEDDING_SKIP_INIT"] = "1"
+    os.environ["ONEFLOW_KERNEL_GLU_ENABLE_DUAL_GEMM_IMPL"] = "0"
+    os.environ["ONEFLOW_MLIR_GROUP_MATMUL_QUANT"] = "1"
+    os.environ["ONEFLOW_FUSE_QUANT_TO_MATMUL"] = "0"
+    # os.environ["ONEFLOW_MLIR_FUSE_KERNEL_LAUNCH"] = "1"
+    # os.environ["ONEFLOW_KERNEL_ENABLE_CUDA_GRAPH"] = "1"
+    
 class QuantDiffusionPipeline:
     def __init__(
         self,
